+++ src/components/focuser/settings/setFocuserReverse.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    v-if="store.focuserInfo.CanReverse"
+    v-if="!store.focuserInfo.CanReverse"
     class="flex items-center justify-between border border-gray-500 p-2 rounded-lg"
   >
     <span class="text-sm font-medium text-gray-300">
@@ -15,7 +15,7 @@
 //#################################
 import { ref, onMounted } from 'vue';
 import { apiStore } from '@/store/store';
-import apiService from '@/services/apiService';
+import apiPinsService from '@/services/apiPinsService';
 import toggleButton from '@/components/helpers/toggleButton.vue';
 
 const store = apiStore();
@@ -24,10 +24,9 @@ const isEnabled = ref(false);
 async function updateSetting() {
   isEnabled.value = !isEnabled.value;
   try {
-    const response = await apiService.profileChangeValue(
-      'FocuserSettings-Reverse',
-      isEnabled.value
-    );
+    const response = await apiPinsService.focuserAction(`reverse?value=${isEnabled.value}`);
+    await apiService.profileChangeValue('FocuserSettings-Reverse', isEnabled.value);
+
     if (!response.Success) {
       // Revert on error
       isEnabled.value = !isEnabled.value;